ui view updated , enemyName and enemyType enums
- take dmg function in health component can take in float now instead of
ints
- ui view updated
- wavetime manager milliseconds counting
- added enum type enemyName that is the name of the enemy and enum type
witch is the type of the enemy : fast , tanky , boss , flying..etc
coin drops increase by each wave (wave +1) at the moment

+++ core/src/com/mygdx/game/entites/entitiycomponents/HealthComponent.java
@@ -3,7 +3,7 @@ package com.mygdx.game.entites.entitiycomponents;
 import com.badlogic.ashley.core.Component;
 
 public class HealthComponent implements Component {
-	public int health = 0;
+	public float health = 0;
 	public int maxHealth = 0;
 	public boolean isDead = false;
 
@@ -12,7 +12,7 @@ public class HealthComponent implements Component {
 		this.maxHealth = maxHealth;
 	}
 	
-	public void takeDamage(int dmg){
+	public void takeDamage(float dmg){
 		health -= dmg;
 		if (health <= 0) isDead = true;
 	}
